Fixing small errors from last commit
Further testing found errors in those files, they now work properly,
added basic tests to tests.sql (must be verified visually)

+++ sql/fn_get_user_photos.sql
@@ -9,7 +9,7 @@ RETURNS TABLE
 	(imagePath 		VARCHAR(250), 
 	description 	VARCHAR(100), 
 	uploadDate 		TIMESTAMP,
-	categoryID		BIGINT,
+	categoryID		INT,
 	shortName 		VARCHAR(25)) AS $$
 BEGIN
     RETURN QUERY
